+++ HttpClientUtilities.Tests/FluentUriBuilderTests.cs
@@ -123,7 +123,7 @@ namespace HttpClientUtilities.Tests
         [MemberData(nameof(ItShouldBuildAQueryStringData))]
         public void ItShouldBuildAUri(FluentUriBuilder builder, string expected)
         {
-            ((string)builder).Should().Be(expected);
+            builder.ToString().Should().Be(expected);
             ((Uri)builder).Should().Be(new Uri(expected, UriKind.Relative));
             builder.ToUri(new Uri("https://example.org/")).Should().Be(new Uri(new Uri("https://example.org/"), expected));
         }