+++ HttpClientUtilities.Tests/Server/SampleModel.cs
@@ -1,0 +1,9 @@
+﻿namespace HttpClientUtilities.Tests.Server
+{
+    public class SampleModel
+    {
+        public const string DefaultName = "Foo";
+
+        public string Name { get; set; } = DefaultName;
+    }
+}